+++ github-updater-additions.php
@@ -13,7 +13,7 @@
  * Plugin Name:       GitHub Updater Additions
  * Plugin URI:        https://github.com/afragen/github-updater-additions
  * Description:       Add installed repositories lacking required headers to the GitHub Updater plugin.
- * Version:           4.0.2
+ * Version:           4.0.2.1
  * Author:            Andy Fragen
  * License:           MIT
  * Network:           true