+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* only add appropriate repo type to GitHub Updater
 
 #### 4.0.2 / 2020-02-05
 * ensure GitHub Updater is running